requirements.txt
streamlit
python-dotenv
google-generativeai
langchain
pypdf2
langchain-google-genai
